# apps/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Lelang Ikan API",
        "description": "API untuk aplikasi e-lelang ikan realtime",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local API Base"
        }
    ],
    "paths": {
        "/admin/analytics/summary": {
            "get": {
                "tags": [
                    "Admin - Analytics"
                ],
                "summary": "Dashboard summary KPI (admin)",
                "operationId": "bb95524a9a2c6734782d1735e942fafa",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/analytics/time-series": {
            "get": {
                "tags": [
                    "Admin - Analytics"
                ],
                "summary": "Orders & GMV time-series (admin)",
                "operationId": "1938c3c90aea8b652ee79bd96c5b0ed0",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/reports/orders.csv": {
            "get": {
                "tags": [
                    "Admin - Reports"
                ],
                "summary": "Export orders CSV (admin)",
                "operationId": "aba85d11e01657a450aa0b02848468c1",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-02-01"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2026-02-25"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auctions/{id}/bids": {
            "get": {
                "tags": [
                    "Bids"
                ],
                "summary": "List bids for an auction (public)",
                "operationId": "19a8b08ffe876608dbea5c08255f5781",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (Laravel paginator)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page. Max is capped by server.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Bids"
                ],
                "summary": "Place a bid (buyer only, auction must be live)",
                "operationId": "3a0f9bca4049f64650cee0e6e8363265",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "example": 150000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auctions": {
            "get": {
                "tags": [
                    "Auctions"
                ],
                "summary": "Public list auctions (filter by status)",
                "operationId": "07df2a586726be0820025c39b2bb56a9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "scheduled",
                                "live",
                                "ended"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (Laravel paginator)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page. Max is capped by server.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auctions/{id}": {
            "get": {
                "tags": [
                    "Auctions"
                ],
                "summary": "Public auction detail",
                "operationId": "592df8c1ed2290138ce07137b2dac599",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/commodities": {
            "get": {
                "tags": [
                    "Commodities"
                ],
                "summary": "Public list commodities (published)",
                "operationId": "2cddc8966b30b245e5121903301dddc3",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (Laravel paginator)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page. Max is capped by server.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/commodities/{id}": {
            "get": {
                "tags": [
                    "Commodities"
                ],
                "summary": "Public detail commodity (published)",
                "operationId": "fda0269f15af878b7b0afed8c679d37b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan Bearer token dari endpoint login/register",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}